Security Update: httpclient library to latest version (2.5.1)
This is a security patch. With the latest version of httpclient, if you
use this gem in an app already using httpclient, you will still be able
to apply the security patch referenced here:

https://github.com/nahi/httpclient/issues/202#issuecomment-59154062

Also removed Gemfile.lock and added the file to .gitignore. It is not
good to check in Gemfile.lock into version control, since it enforces
precision that does not exist in the gem command, which is used to
install gems in practice.

See http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/